+++ examples/resources/nxos_feature_evpn/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_feature_evpn.example "sys/fm/evpn"