+++ examples/resources/nxos_feature_evpn/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_feature_evpn" "example" {
+  admin_state = "enabled"
+}