+++ examples/resources/nxos_feature_hsrp/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_feature_hsrp.example "sys/fm/hsrp"