+++ examples/resources/nxos_feature_isis/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_feature_isis" "example" {
+  admin_state = "enabled"
+}