+++ examples/resources/nxos_feature_lacp/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_feature_lacp.example "sys/fm/lacp"