+++ examples/resources/nxos_feature_lacp/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_feature_lacp" "example" {
+  admin_state = "enabled"
+}