+++ examples/resources/nxos_feature_ptp/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_feature_ptp.example "sys/fm/ptp"