+++ examples/resources/nxos_feature_ptp/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_feature_ptp" "example" {
+  admin_state = "enabled"
+}