+++ examples/resources/nxos_feature_pvlan/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_feature_pvlan.example "sys/fm/pvlan"