+++ examples/resources/nxos_feature_ssh/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_feature_ssh.example "sys/fm/ssh"