+++ examples/resources/nxos_feature_tacacs/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_feature_tacacs.example "sys/fm/tacacsplus"