+++ examples/resources/nxos_feature_vpc/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_feature_vpc.example "sys/fm/vpc"