+++ examples/resources/nxos_feature_vn_segment/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_feature_vn_segment.example "sys/fm/vnsegment"