+++ examples/resources/nxos_feature_vn_segment/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_feature_vn_segment" "example" {
+  admin_state = "enabled"
+}