+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add nxos_feature_bgp resource and data source
 - Add nxos_feature_evpn resource and data source
 - Add nxos_feature_hsrp resource and data source
+- Add nxos_feature_isis resource and data source
 
 ## 0.3.0