+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add nxos_feature_ospfv3 resource and data source
 - Add nxos_feature_ptp resource and data source
 - Add nxos_feature_pvlan resource and data source
+- Add nxos_feature_ssh resource and data source
 
 ## 0.3.0